Added oath creds - DO NOT USE
Just for a test. Just need to add the config value for your personal access token in a config "githubPersonalAccessToken": "XXX" and pull it in. This is far form done. But the idea works.. The goal is to make a Launch in Colab. But first you need a gist then it's straight forward.